refactor: move CodeSignature back into vexide_startup
