+++ packages/vexide-startup/src/lib.rs
@@ -13,6 +13,72 @@
 #![feature(asm_experimental_arch)]
 #![allow(clippy::needless_doctest_main)]
 
+use bitflags::bitflags;
+
+/// Identifies the type of binary to VEXos.
+#[repr(u32)]
+#[non_exhaustive]
+pub enum ProgramType {
+    /// User program binary.
+    User = 0,
+}
+
+/// The owner (originator) of the user program
+#[repr(u32)]
+pub enum ProgramOwner {
+    /// Program is a system binary.
+    System = 0,
+
+    /// Program originated from VEX.
+    Vex = 1,
+
+    /// Program originated from a partner developer.
+    Partner = 2,
+}
+
+bitflags! {
+    /// Program Flags
+    ///
+    /// These bitflags are part of the [`CodeSignature`] that determine some small
+    /// aspects of program behavior when running under VEXos. This struct contains
+    /// the flags with publicly documented behavior.
+    #[derive(Default, Debug, Clone, Copy, Eq, PartialEq)]
+    pub struct ProgramFlags: u32 {
+        /// Default graphics colors (foreground/background) will be inverted.
+        const INVERT_DEFAULT_GRAPHICS = 1 << 0;
+
+        /// VEXos scheduler simple tasks will be killed when the program requests exit.
+        const KILL_TASKS_ON_EXIT = 1 << 1;
+
+        /// Default graphics colors (foreground/background) will invert based on
+        /// the selected system theme.
+        const THEMED_DEFAULT_GRAPHICS = 1 << 2;
+    }
+}
+
+/// Program Code Signature
+///
+/// The first 16 bytes of a VEX user code binary contain a user code signature,
+/// containing some basic metadata and startup flags about the program. This
+/// signature must be at the start of the binary for booting to occur.
+#[derive(Default, Debug, Clone, Copy, Eq, PartialEq)]
+pub struct CodeSignature(vex_sdk::vcodesig, [u32; 4]);
+
+impl CodeSignature {
+    /// Creates a new signature given a program type, owner, and flags.
+    pub const fn new(program_type: ProgramType, owner: ProgramOwner, flags: ProgramFlags) -> Self {
+        Self(
+            vex_sdk::vcodesig {
+                magic: vex_sdk::V5_SIG_MAGIC,
+                r#type: program_type as _,
+                owner: owner as _,
+                options: flags.bits(),
+            },
+            [0; 4],
+        )
+    }
+}
+
 extern "C" {
     // These symbols don't have real types so this is a little bit of a hack
     static mut __bss_start: u32;
